+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- dive list: make dive guide visible in dive list [#3382]
 - general: rename dive master to dive guide
 - libdivecomputer: add support for latest BLE hardware in OSTC dive computers